Optimize API requests and refactor the shit out of event-list
